Remove language nb from Makefile
Before this commit, it was creating two same files for Norwegian Bokmål
language. When they are present both files, the plugin can not be added
to Weblate.

While looking at this website:
http://people.skolelinux.org/pere/blog/Spr_kkoder_for_POSIX_locale_i_Norge.html

I see that only nb_NO should be used and nb should not be used, which
was confirmed by Norwegian guy.